Fix issue 679
https://code.google.com/p/singularity-viewer/issues/detail?id=679

Instead of making the cursor just always an arrow, I changed the
'working' cursor from a watch to an arrow + hour glass. Note that the
BMP of the 'wait' cursor was exactly the same as the 'working' cursor,
so if anywhere 'working' should NOT have an arrow, then that should be
replaced with the UI_CURSOR_WAIT to restore the old cursor.

Obviously, the blinking is fixed too - cause that is what this all was
about.