Fix crash when granting/removing build rights to others.
Bug was introduced in c701c615664f598818d070775b1509103a24f03e
for single persons, but already existed for when granting rights
to more than one person at a time.

There is still a problem left now that you have to confirm
several times (once, but also had to confirm two or three times)
before it gets through. That doesn't seem related to THIS bug,
but seems to be a new bug, possibly introduced in the same commit
mentioned above.