Fix Issue 459: Appearance, Create Outfit does not keep edits to outfit - HAPPENS EVERY TIME
Automatically detach copyable attachments early, so that changes will be kept (LLMakeOutfitDialog was changed to update the outfit correctly, despite items being removed. It should probably be cleaned up further)
Adds an extra checkbox to make outfit dialog to disable this new auto-detach behavior, for users who rely on the old behavior.

Also fixes a bug I noticed in LLMakeOutfitDialog in which the use links checkbox would be enabled even if the grid doesn't support links
Had to reposition everything to get the new checkbox to sit nicely, but in doing so changed bottoms to bottom_deltas so it should be easier in the future, layout remains the same, of course, shouldn't be noticed.