Fix crash when previewing an animation.
When uploading an animation, playing the to-be-uploaded
animation on ones own avatar while ground sitting would
crash the viewer.

This fixes that, and makes the avatar cleanly stand up first.
Rationale: the alternative is to make the user learn by
means of crashing, being frozen or just not seeing the right
animation that they have to stand before uploading an animation.
There is no reason not to automate that.