Feature Request: Allow typing login uris into the grids combo box on the login screen
Furthermore, users may now type in grid name, grid nick, or grid login uri into the box and it will resolve to an existing grid they've used before, or (for uris only) try to add the new grid with the data it provides.
Input will have trailing and leading spaces trimmed.

Thanks to Diva Canto for this brilliant request!